fixed typo in package
